refactor: modularize popup architecture and implement smart assistant system
- Split popup into specialized modules (PopupController, UIManager, DebugTools, DataExporter)
- Add complete smart assistant framework with execution engine and messaging service
- Implement style injection system for dynamic UI enhancements
- Add feedback collection system for user interaction analytics
- Improve code organization with better separation of concerns
- Remove obsolete patch files and reorganize component structure
