--- popup-refactor.patch
@@ -1,4 +1,0 @@
-diff --git a/src/popup.ts b/src/popup/index.ts
-similarity index 100%
-rename from src/popup.ts
-rename to src/popup/index.ts
